modify the EditForm with a conditonal fetch, dynamit for title and button, dynamic fetch method, and controlled inputs
