+++ app/javascript/components/RecipeForm.jsx
@@ -1,11 +1,25 @@
-import React, { useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import { Link, useNavigate, useParams } from "react-router-dom";
 
-const NewRecipe = () => {
+const RecipeForm = () => {
+  const params = useParams();
   const navigate = useNavigate();
-  const [name, setName] = useState("");
-  const [ingredients, setIngredients] = useState("");
-  const [instruction, setInstruction] = useState("");
+  const [recipe, setRecipe] = useState({ name: "", ingredients: "", instruction: "" });
+
+  useEffect(() => {
+    if (params.id) {
+      const url = `/api/v1/recipes/${params.id}/edit`;
+      fetch(url)
+        .then((response) => {
+          if (response.ok) {
+            return response.json();
+          }
+          throw new Error("Network response was not ok.");
+        })
+        .then((response) => setRecipe(response))
+        .catch(() => navigate("/recipes"));
+    }
+  }, [params.id, navigate]);
 
   const stripHtmlEntities = (str) => {
     return String(str)
@@ -14,16 +28,30 @@ const NewRecipe = () => {
       .replace(/>/g, "&gt;");
   };
 
-  const onChange = (event, setFunction) => {
-    setFunction(event.target.value);
+  const updateRecipe = (key, value) => {
+    setRecipe((previousRecipe) => ({ ...previousRecipe, [key]: value }));
+  };
+
+  const handleInputChange = (e) => {
+    const { target } = e;
+    const { name } = target;
+    const value = target.value;
+
+    updateRecipe(name, value);
   };
 
   const onSubmit = (event) => {
     event.preventDefault();
-    const url = "/api/v1/recipes";
+    const id = params.id;
+    const url = id ? `/api/v1/recipes/${id}` : `/api/v1/recipes`;
 
-    if (name.length === 0 || ingredients.length === 0 || instruction.length === 0)
+    const name = recipe.name;
+    const ingredients = recipe.ingredients;
+    const instruction = recipe.instruction;
+
+    if (name.length === 0 || ingredients.length === 0 || instruction.length === 0) {
       return;
+    }
 
     const body = {
       name,
@@ -31,9 +59,11 @@ const NewRecipe = () => {
       instruction: stripHtmlEntities(instruction),
     };
 
+    const method = id ? "PUT" : "POST";
     const token = document.querySelector('meta[name="csrf-token"]').content;
+
     fetch(url, {
-      method: "POST",
+      method: method,
       headers: {
         "X-CSRF-Token": token,
         "Content-Type": "application/json",
@@ -55,9 +85,9 @@ const NewRecipe = () => {
       <div className="row">
         <div className="col-sm-12 col-lg-6 offset-lg-3">
           <h1 className="font-weight-normal mb-5">
-            Add a new recipe to our awesome recipe collection.
+            {params.id ? window.EDIT_RECIPE_FORM_TITLE_TEXT : window.CREATE_NEW_RECIPE_FORM_TEXT}
           </h1>
-          <form id="recipeForm" onSubmit={onSubmit}>
+          <form onSubmit={onSubmit}>
             <div className="form-group">
               <label htmlFor="recipeName">{window.RECIPE_FORM_NAME_FIELD}</label>
               <input
@@ -66,7 +96,8 @@ const NewRecipe = () => {
                 id="recipeName"
                 className="form-control"
                 required
-                onChange={(event) => onChange(event, setName)}
+                value={recipe.name}
+                onChange={handleInputChange}
               />
             </div>
             <div className="form-group">
@@ -77,7 +108,8 @@ const NewRecipe = () => {
                 id="recipeIngredients"
                 className="form-control"
                 required
-                onChange={(event) => onChange(event, setIngredients)}
+                value={recipe.ingredients}
+                onChange={handleInputChange}
               />
               <small id="ingredientsHelp" className="form-text text-muted">
                 Separate each ingredient with a comma.
@@ -90,10 +122,11 @@ const NewRecipe = () => {
               name="instruction"
               rows="5"
               required
-              onChange={(event) => onChange(event, setInstruction)}
+              value={recipe.instruction}
+              onChange={handleInputChange}
             />
-            <button id="submit" type="submit" className="btn custom-button mt-3">
-              {window.CREATE_RECIPE_BUTTON_TEXT}
+            <button type="submit" className="btn custom-button mt-3">
+              {params.id ? window.UPDATE_RECIPE_BUTTON_TEXT : window.CREATE_RECIPE_BUTTON_TEXT}
             </button>
             <Link to="/recipes" className="btn btn-link mt-3">
               {window.BACK_TO_RECIPES_BUTTON_TEXT}
@@ -105,4 +138,4 @@ const NewRecipe = () => {
   );
 };
 
-export default NewRecipe;
+export default RecipeForm;
