+++ app/views/layouts/application.html.erb
@@ -13,9 +13,11 @@
   <body>
     <script>window.BACK_TO_RECIPES_BUTTON_TEXT = "<%= j(BACK_TO_RECIPES_BUTTON_TEXT) %>";</script>
     <script>window.CREATE_NEW_RECIPE_BUTTON_TEXT = "<%= j(CREATE_NEW_RECIPE_BUTTON_TEXT) %>";</script>
+    <script>window.CREATE_NEW_RECIPE_FORM_TEXT = "<%= j(CREATE_NEW_RECIPE_FORM_TEXT) %>";</script>
     <script>window.CREATE_RECIPE_BUTTON_TEXT = "<%= j(CREATE_RECIPE_BUTTON_TEXT) %>";</script>
     <script>window.DELETE_RECIPE_BUTTON_TEXT = "<%= j(DELETE_RECIPE_BUTTON_TEXT) %>";</script>
     <script>window.EDIT_RECIPE_BUTTON_TEXT = "<%= j(EDIT_RECIPE_BUTTON_TEXT) %>";</script>
+    <script>window.EDIT_RECIPE_FORM_TITLE_TEXT = "<%= j(EDIT_RECIPE_FORM_TITLE_TEXT) %>";</script>
     <script>window.HOME_BUTTON_TEXT = "<%= j(HOME_BUTTON_TEXT) %>";</script>
     <script>window.HOME_HEADER_TEXT = "<%= j(HOME_HEADER_TEXT) %>";</script>
     <script>window.RECIPES_INDEX_HEADER_TEXT = "<%= j(RECIPES_INDEX_HEADER_TEXT) %>";</script>
