+++ config/initializers/my_constants.rb
@@ -1,8 +1,10 @@
 BACK_TO_RECIPES_BUTTON_TEXT = "Back To Recipes"
 CREATE_NEW_RECIPE_BUTTON_TEXT = "Create New Recipe"
+CREATE_NEW_RECIPE_FORM_TEXT = "Create New Recipe"
 CREATE_RECIPE_BUTTON_TEXT = "Create Recipe"
 DELETE_RECIPE_BUTTON_TEXT = "Delete Recipe"
 EDIT_RECIPE_BUTTON_TEXT = "Edit Recipe"
+EDIT_RECIPE_FORM_TITLE_TEXT = "Edit This Recipe"
 HOME_BUTTON_TEXT = "Home"
 HOME_HEADER_TEXT = "Food Recipes"
 RECIPES_INDEX_HEADER_TEXT = "Recipes for every occasion!"
